--- src/training/steps/pre_training/feature_generation_enhanced_m1_optimized_step.py
@@ -1,797 +1,0 @@
-"""
-Enhanced M1-Optimized Feature Generation Step
-
-This step implements comprehensive M1 hardware optimization for feature generation,
-utilizing all available M1 capabilities including Neural Engine, unified memory,
-advanced CPU optimization, and enhanced GPU acceleration.
-"""
-
-from __future__ import annotations
-
-import logging
-import json
-import warnings
-import pandas as pd
-import numpy as np
-import copy
-import asyncio
-import time
-import gc
-import os
-from concurrent.futures import ThreadPoolExecutor
-from threading import Lock
-from typing import Any, Dict, List, Optional, Tuple
-from datetime import datetime
-from pathlib import Path
-from dataclasses import dataclass, field
-
-from src.training.steps.base_step import BaseStep
-from src.utils.common_operations import safe_dataframe_operation
-from src.utils.matrix_operations import safe_matrix_multiply, optimize_dataframe
-
-# Enhanced M1 hardware optimization imports
-from src.utils.hardware import (
-    # Core M1 optimizers
-    get_comprehensive_optimizer, M1ComprehensiveOptimizer, ComprehensiveConfig,
-    get_unified_memory_manager, M1UnifiedMemoryManager,
-    get_advanced_cpu_optimizer, M1AdvancedCPUOptimizer,
-    get_enhanced_gpu_manager, EnhancedM1GPUManager,
-    get_neural_engine_manager, M1NeuralEngineManager,
-    get_advanced_memory_manager, AdvancedMemoryManager,
-    get_integrated_hardware_manager, IntegratedHardwareConfig,
-    
-    # Optimization decorators
-    memory_optimized, gc_optimized, chunked_processing_auto,
-    comprehensive_memory_optimization, MemoryOptimizationLevel,
-    auto_optimize, performance_tracked, smart_cache,
-    optimize_dataframe_default, optimize_numpy_array_default,
-    
-    # Workload types and optimization levels
-    WorkloadType, OptimizationLevel, OptimizationStrategy, WorkloadCategory,
-    get_memory_optimization_stats, force_cleanup
-)
-
-# Import tprint utilities
-try:
-    from src.utils.tprint import (
-        tprint, tprint_info, tprint_success, tprint_warning, tprint_error, tprint_debug
-    )
-except ImportError:
-    def tprint(*args, **kwargs): print(*args)
-    def tprint_info(*args, **kwargs): print("INFO:", *args)
-    def tprint_success(*args, **kwargs): print("SUCCESS:", *args)
-    def tprint_warning(*args, **kwargs): print("WARNING:", *args)
-    def tprint_error(*args, **kwargs): print("ERROR:", *args)
-    def tprint_debug(*args, **kwargs): print("DEBUG:", *args)
-
-# Enhanced M1 optimization configuration
-@dataclass
-class M1FeatureGenerationConfig:
-    """M1-optimized feature generation configuration."""
-    # M1 optimization settings
-    enable_neural_engine: bool = True
-    enable_unified_memory: bool = True
-    enable_advanced_cpu: bool = True
-    enable_enhanced_gpu: bool = True
-    enable_adaptive_optimization: bool = True
-    
-    # Performance settings
-    optimization_strategy: OptimizationStrategy = OptimizationStrategy.MAXIMUM_PERFORMANCE
-    workload_category: WorkloadCategory = WorkloadCategory.FINANCIAL_MODELING
-    
-    # Memory settings
-    memory_optimization_level: MemoryOptimizationLevel = MemoryOptimizationLevel.MAXIMUM
-    enable_memory_pooling: bool = True
-    enable_dynamic_allocation: bool = True
-    
-    # Processing settings
-    enable_parallel_processing: bool = True
-    max_parallel_workers: int = 8  # Optimized for M1
-    enable_chunked_processing: bool = True
-    chunk_size_mb: float = 100.0
-    
-    # Caching settings
-    enable_intelligent_caching: bool = True
-    cache_ttl: float = 3600.0
-    enable_feature_caching: bool = True
-
-@dataclass
-class M1FeatureGenerationResult:
-    """M1-optimized feature generation result."""
-    feature_names: List[str]
-    feature_data: pd.DataFrame
-    generated_features: pd.DataFrame
-    generation_time: float
-    n_features_generated: int
-    cache_hit: bool
-    memory_usage_mb: float
-    success: bool
-    feature_categories: List[str] = field(default_factory=list)
-    error_message: Optional[str] = None
-    metadata: Dict[str, Any] = field(default_factory=dict)
-    optimization_stats: Dict[str, Any] = field(default_factory=dict)
-    generation_metrics: Dict[str, Any] = field(default_factory=dict)
-    artifacts: Dict[str, Any] = field(default_factory=dict)
-    
-    # M1-specific metrics
-    neural_engine_utilization: float = 0.0
-    gpu_utilization: float = 0.0
-    cpu_utilization: float = 0.0
-    memory_efficiency: float = 0.0
-    optimization_applied: List[str] = field(default_factory=list)
-
-    @property
-    def features(self) -> pd.DataFrame:
-        """Backward-compatible accessor."""
-        return self.generated_features
-
-class M1OptimizedFeatureGenerationStep(BaseStep):
-    """M1-optimized feature generation step with comprehensive hardware utilization."""
-
-    def __init__(self, config: Optional[Dict[str, Any]] = None):
-        """Initialize the M1-optimized feature generation step."""
-        super().__init__("m1_optimized_feature_generation_step", config)
-        
-        # M1 optimization configuration
-        self.m1_config = M1FeatureGenerationConfig()
-        
-        # Initialize M1 hardware components
-        self._initialize_m1_components()
-        
-        # Performance tracking
-        self.performance_stats = {
-            'total_processing_time': 0.0,
-            'neural_engine_operations': 0,
-            'gpu_accelerations': 0,
-            'cpu_optimizations': 0,
-            'memory_optimizations': 0,
-            'cache_hits': 0,
-            'memory_savings_mb': 0.0,
-            'optimization_applied': []
-        }
-        
-        tprint_success("🚀 M1-Optimized Feature Generation Step initialized")
-
-    def _initialize_m1_components(self):
-        """Initialize all M1 hardware optimization components."""
-        try:
-            tprint_info("🔧 Initializing M1 hardware optimization components")
-            
-            # Initialize M1 Comprehensive Optimizer
-            comprehensive_config = ComprehensiveConfig(
-                optimization_strategy=self.m1_config.optimization_strategy,
-                workload_category=self.m1_config.workload_category,
-                enable_adaptive_optimization=self.m1_config.enable_adaptive_optimization,
-                enable_cross_component_optimization=True,
-                enable_thermal_management=True,
-                enable_power_management=True,
-                enable_comprehensive_monitoring=True,
-                enable_auto_tuning=True
-            )
-            self.comprehensive_optimizer = get_comprehensive_optimizer(comprehensive_config)
-            
-            # Initialize M1 Unified Memory Manager
-            if self.m1_config.enable_unified_memory:
-                self.unified_memory_manager = get_unified_memory_manager()
-                tprint_info("✅ M1 Unified Memory Manager initialized")
-            
-            # Initialize M1 Advanced CPU Optimizer
-            if self.m1_config.enable_advanced_cpu:
-                self.cpu_optimizer = get_advanced_cpu_optimizer()
-                self.cpu_optimizer.optimize_for_workload(WorkloadType.FEATURE_ENGINEERING)
-                tprint_info("✅ M1 Advanced CPU Optimizer initialized")
-            
-            # Initialize M1 Enhanced GPU Manager
-            if self.m1_config.enable_enhanced_gpu:
-                self.gpu_manager = get_enhanced_gpu_manager()
-                if self.gpu_manager.is_available():
-                    tprint_info("✅ M1 Enhanced GPU Manager initialized")
-                else:
-                    tprint_warning("⚠️ M1 GPU not available")
-            
-            # Initialize M1 Neural Engine Manager
-            if self.m1_config.enable_neural_engine:
-                self.neural_engine_manager = get_neural_engine_manager()
-                if self.neural_engine_manager.is_available():
-                    tprint_info("✅ M1 Neural Engine Manager initialized")
-                else:
-                    tprint_warning("⚠️ M1 Neural Engine not available")
-            
-            # Initialize Advanced Memory Manager
-            self.memory_manager = get_advanced_memory_manager()
-            
-            # Initialize Integrated Hardware Manager
-            integrated_config = IntegratedHardwareConfig(
-                memory_limit_gb=16.0,  # Increased for M1
-                enable_automatic_optimization=True,
-                enable_caching=self.m1_config.enable_intelligent_caching,
-                enable_memory_monitoring=True,
-                enable_performance_tracking=True,
-                default_optimization_level=OptimizationLevel.AGGRESSIVE
-            )
-            self.integrated_hardware_manager = get_integrated_hardware_manager(integrated_config)
-            
-            tprint_success("🚀 All M1 hardware components initialized successfully")
-            
-        except Exception as e:
-            tprint_error(f"❌ Failed to initialize M1 components: {e}")
-            self.logger.error(f"M1 component initialization failed: {e}")
-            # Set components to None for graceful degradation
-            self.comprehensive_optimizer = None
-            self.unified_memory_manager = None
-            self.cpu_optimizer = None
-            self.gpu_manager = None
-            self.neural_engine_manager = None
-            self.memory_manager = None
-            self.integrated_hardware_manager = None
-
-    async def execute(self, config: Dict[str, Any]) -> Dict[str, Any]:
-        """Execute M1-optimized feature generation step."""
-        start_time = time.time()
-        self.logger.info("Starting M1-optimized feature generation step")
-        
-        # Extract parameters from config
-        data = config.get('data')
-        targets = config.get('targets')
-        symbol = config.get('symbol', 'ETHUSDT')
-        timeframe = config.get('timeframe', '15m')
-        direction = config.get('direction', 'longs')
-        intensity = config.get('intensity')
-        lookback_days = config.get('lookback_days')
-        start_date = config.get('start_date')
-        end_date = config.get('end_date')
-        exchange = config.get('exchange', 'binance')
-        custom_overrides = config.get('custom_overrides')
-        pipeline_state = config.get('pipeline_state', {})
-        
-        # Set context for enhanced file naming
-        self._set_context(symbol=symbol, exchange=exchange, direction=direction, model='M1Optimized')
-        
-        try:
-            # Validate input data
-            if data is None or len(data) == 0:
-                raise ValueError("Input data is None or empty")
-            
-            # Apply M1 comprehensive optimization to input data
-            optimized_data = await self._apply_m1_comprehensive_optimization(data)
-            
-            # Perform M1-optimized feature generation
-            generation_result = await self._perform_m1_optimized_feature_generation(
-                optimized_data, symbol, timeframe, direction, custom_overrides, targets
-            )
-            
-            if generation_result.success:
-                # Update performance statistics
-                end_time = time.time()
-                self.performance_stats['total_processing_time'] = end_time - start_time
-                
-                # Add M1 optimization statistics
-                generation_result.optimization_stats.update({
-                    'm1_optimizations': {
-                        'neural_engine_utilization': generation_result.neural_engine_utilization,
-                        'gpu_utilization': generation_result.gpu_utilization,
-                        'cpu_utilization': generation_result.cpu_utilization,
-                        'memory_efficiency': generation_result.memory_efficiency,
-                        'optimization_applied': generation_result.optimization_applied,
-                        'total_processing_time': self.performance_stats['total_processing_time'],
-                        'neural_engine_operations': self.performance_stats['neural_engine_operations'],
-                        'gpu_accelerations': self.performance_stats['gpu_accelerations'],
-                        'cpu_optimizations': self.performance_stats['cpu_optimizations'],
-                        'memory_optimizations': self.performance_stats['memory_optimizations'],
-                        'cache_hits': self.performance_stats['cache_hits'],
-                        'memory_savings_mb': self.performance_stats['memory_savings_mb']
-                    }
-                })
-                
-                self.logger.info(f"M1-optimized feature generation completed successfully")
-                self.logger.info(f"Generated {len(generation_result.generated_features.columns)} features")
-                self.logger.info(f"Neural Engine utilization: {generation_result.neural_engine_utilization:.1f}%")
-                self.logger.info(f"GPU utilization: {generation_result.gpu_utilization:.1f}%")
-                self.logger.info(f"Memory efficiency: {generation_result.memory_efficiency:.1f}%")
-                
-                # Save artifacts using BaseStep methods
-                self._save_dataframe(generation_result.generated_features, 'generated_features')
-                self._save_metadata(generation_result.feature_names, 'feature_names')
-                self._save_metadata(generation_result.feature_categories, 'feature_categories')
-                self._save_metadata(generation_result.optimization_stats, 'optimization_stats')
-                
-                # Generate M1 optimization report
-                report_path = await self._generate_m1_optimization_report(
-                    generation_result, symbol, timeframe, direction, exchange
-                )
-                self.logger.info(f"📊 M1 optimization report generated: {report_path}")
-            else:
-                self.logger.error(f"M1-optimized feature generation failed: {generation_result.error_message}")
-
-            return generation_result
-
-        except Exception as e:
-            self.logger.error(f"M1-optimized feature generation step failed: {e}")
-            return M1FeatureGenerationResult(
-                feature_names=[],
-                feature_data=pd.DataFrame(),
-                generated_features=pd.DataFrame(),
-                feature_categories=[],
-                generation_time=0.0,
-                n_features_generated=0,
-                cache_hit=False,
-                memory_usage_mb=0.0,
-                success=False,
-                error_message=str(e),
-                metadata={},
-                optimization_stats={}
-            )
-        
-        finally:
-            # Cleanup M1 resources
-            await self._cleanup_m1_resources()
-
-    @comprehensive_memory_optimization(
-        optimization_level=MemoryOptimizationLevel.MAXIMUM,
-        enable_caching=True,
-        enable_chunking=True,
-        enable_gc=True,
-        enable_pools=True
-    )
-    async def _apply_m1_comprehensive_optimization(self, data: pd.DataFrame) -> pd.DataFrame:
-        """Apply comprehensive M1 optimization to input data."""
-        try:
-            tprint_info("🚀 Applying M1 comprehensive optimization to input data")
-            
-            if not isinstance(data, pd.DataFrame) or data.empty:
-                return data
-            
-            initial_memory = data.memory_usage(deep=True).sum()
-            
-            # Use M1 Comprehensive Optimizer
-            if self.comprehensive_optimizer:
-                optimized_data = self.comprehensive_optimizer.optimize_dataframe(
-                    data, 
-                    workload_type=WorkloadType.FEATURE_ENGINEERING,
-                    optimization_strategy=self.m1_config.optimization_strategy
-                )
-                self.performance_stats['cpu_optimizations'] += 1
-                tprint_info("✅ M1 Comprehensive Optimizer applied")
-            else:
-                optimized_data = data
-            
-            # Apply M1 Unified Memory optimization
-            if self.unified_memory_manager:
-                optimized_data = self.unified_memory_manager.optimize_dataframe(optimized_data)
-                self.performance_stats['memory_optimizations'] += 1
-                tprint_info("✅ M1 Unified Memory optimization applied")
-            
-            # Apply M1 Advanced CPU optimization
-            if self.cpu_optimizer:
-                optimized_data = self.cpu_optimizer.optimize_dataframe(optimized_data)
-                self.performance_stats['cpu_optimizations'] += 1
-                tprint_info("✅ M1 Advanced CPU optimization applied")
-            
-            # Apply M1 Enhanced GPU optimization
-            if self.gpu_manager and self.gpu_manager.is_available():
-                optimized_data = self.gpu_manager.optimize_dataframe(optimized_data)
-                self.performance_stats['gpu_accelerations'] += 1
-                tprint_info("✅ M1 Enhanced GPU optimization applied")
-            
-            # Apply Advanced Memory Manager optimization
-            if self.memory_manager:
-                optimized_data = self.memory_manager.optimize_dataframe(optimized_data)
-                self.performance_stats['memory_optimizations'] += 1
-            
-            # Calculate memory savings
-            final_memory = optimized_data.memory_usage(deep=True).sum()
-            memory_saved = initial_memory - final_memory
-            
-            if memory_saved > 0:
-                tprint_success(f"🚀 M1 optimization: {memory_saved / 1024**2:.2f} MB saved")
-                self.performance_stats['memory_savings_mb'] += memory_saved / 1024**2
-            
-            return optimized_data
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ M1 comprehensive optimization failed: {e}")
-            return optimize_dataframe_default(data)
-
-    @memory_optimized(optimization_level=MemoryOptimizationLevel.MAXIMUM)
-    @performance_tracked
-    async def _perform_m1_optimized_feature_generation(self, data: pd.DataFrame, symbol: str,
-                                                      timeframe: str, direction: str,
-                                                      custom_overrides: Optional[Dict[str, Any]],
-                                                      targets: Optional[pd.Series] = None) -> M1FeatureGenerationResult:
-        """Perform M1-optimized feature generation using all available M1 capabilities."""
-        
-        start_time = time.time()
-        
-        try:
-            tprint_info("🚀 Starting M1-optimized feature generation")
-            
-            # Initialize result tracking
-            neural_engine_utilization = 0.0
-            gpu_utilization = 0.0
-            cpu_utilization = 0.0
-            memory_efficiency = 0.0
-            optimization_applied = []
-            
-            # Use M1 Neural Engine for ML feature generation
-            ml_features = pd.DataFrame()
-            if self.neural_engine_manager and self.neural_engine_manager.is_available():
-                try:
-                    tprint_info("🧠 Using M1 Neural Engine for ML feature generation")
-                    ml_features = self.neural_engine_manager.process_ml_features(
-                        data, 
-                        feature_types=['technical_indicators', 'statistical_features', 'pattern_recognition']
-                    )
-                    neural_engine_utilization = 85.0  # Estimated utilization
-                    optimization_applied.append('neural_engine_ml_features')
-                    self.performance_stats['neural_engine_operations'] += 1
-                    tprint_success("✅ M1 Neural Engine ML features generated")
-                except Exception as e:
-                    tprint_warning(f"⚠️ Neural Engine ML features failed: {e}")
-            
-            # Use M1 Enhanced GPU for vectorized operations
-            gpu_features = pd.DataFrame()
-            if self.gpu_manager and self.gpu_manager.is_available():
-                try:
-                    tprint_info("🎮 Using M1 Enhanced GPU for vectorized operations")
-                    gpu_features = self.gpu_manager.accelerate_vectorized_operations(
-                        data,
-                        operations=['rolling_calculations', 'statistical_operations', 'technical_indicators']
-                    )
-                    gpu_utilization = 70.0  # Estimated utilization
-                    optimization_applied.append('gpu_vectorized_operations')
-                    self.performance_stats['gpu_accelerations'] += 1
-                    tprint_success("✅ M1 Enhanced GPU vectorized features generated")
-                except Exception as e:
-                    tprint_warning(f"⚠️ GPU vectorized operations failed: {e}")
-            
-            # Use M1 Advanced CPU for traditional feature generation
-            cpu_features = pd.DataFrame()
-            if self.cpu_optimizer:
-                try:
-                    tprint_info("💻 Using M1 Advanced CPU for traditional features")
-                    cpu_features = self.cpu_optimizer.optimize_feature_generation(
-                        data,
-                        feature_categories=['price_features', 'volume_features', 'volatility_features']
-                    )
-                    cpu_utilization = 80.0  # Estimated utilization
-                    optimization_applied.append('cpu_traditional_features')
-                    self.performance_stats['cpu_optimizations'] += 1
-                    tprint_success("✅ M1 Advanced CPU traditional features generated")
-                except Exception as e:
-                    tprint_warning(f"⚠️ CPU traditional features failed: {e}")
-            
-            # Combine all generated features
-            all_features = []
-            if not ml_features.empty:
-                all_features.append(ml_features)
-            if not gpu_features.empty:
-                all_features.append(gpu_features)
-            if not cpu_features.empty:
-                all_features.append(cpu_features)
-            
-            if all_features:
-                generated_features_df = pd.concat(all_features, axis=1)
-            else:
-                # Fallback to basic feature generation
-                tprint_warning("⚠️ All M1 optimizations failed, using fallback")
-                generated_features_df = self._fallback_feature_generation(data)
-                optimization_applied.append('fallback_generation')
-            
-            # Apply M1 Unified Memory optimization to final result
-            if self.unified_memory_manager:
-                generated_features_df = self.unified_memory_manager.optimize_dataframe(generated_features_df)
-                optimization_applied.append('unified_memory_optimization')
-            
-            # Calculate memory efficiency
-            original_memory = data.memory_usage(deep=True).sum()
-            final_memory = generated_features_df.memory_usage(deep=True).sum()
-            memory_efficiency = (1 - (final_memory / original_memory)) * 100 if original_memory > 0 else 0
-            
-            # Create feature names and categories
-            feature_names = list(generated_features_df.columns)
-            feature_categories = self._categorize_features(feature_names)
-            
-            # Calculate generation metrics
-            generation_duration = time.time() - start_time
-            
-            return M1FeatureGenerationResult(
-                feature_names=feature_names,
-                feature_data=generated_features_df,
-                generated_features=generated_features_df,
-                feature_categories=feature_categories,
-                generation_time=generation_duration,
-                n_features_generated=len(feature_names),
-                cache_hit=False,
-                memory_usage_mb=generated_features_df.memory_usage(deep=True).sum() / 1024 / 1024,
-                success=True,
-                error_message=None,
-                optimization_stats={},
-                generation_metrics={
-                    'generation_time': generation_duration,
-                    'features_generated': len(feature_names),
-                    'memory_usage_mb': generated_features_df.memory_usage(deep=True).sum() / 1024 / 1024,
-                    'categories_count': len(feature_categories)
-                },
-                artifacts={
-                    'feature_dataframe': generated_features_df,
-                    'feature_names': feature_names,
-                    'feature_categories': feature_categories,
-                    'raw_dataframe': data
-                },
-                neural_engine_utilization=neural_engine_utilization,
-                gpu_utilization=gpu_utilization,
-                cpu_utilization=cpu_utilization,
-                memory_efficiency=memory_efficiency,
-                optimization_applied=optimization_applied
-            )
-            
-        except Exception as e:
-            self.logger.error(f"M1-optimized feature generation failed: {e}")
-            return M1FeatureGenerationResult(
-                feature_names=[],
-                feature_data=pd.DataFrame(),
-                generated_features=pd.DataFrame(),
-                feature_categories=[],
-                generation_time=0.0,
-                n_features_generated=0,
-                cache_hit=False,
-                memory_usage_mb=0.0,
-                success=False,
-                error_message=str(e),
-                metadata={},
-                optimization_stats={}
-            )
-
-    def _fallback_feature_generation(self, data: pd.DataFrame) -> pd.DataFrame:
-        """Fallback feature generation when M1 optimizations fail."""
-        try:
-            tprint_info("🔄 Using fallback feature generation")
-            
-            # Basic technical indicators as fallback
-            features = {}
-            
-            if 'close' in data.columns:
-                # Simple moving averages
-                features['sma_20'] = data['close'].rolling(window=20).mean()
-                features['sma_50'] = data['close'].rolling(window=50).mean()
-                
-                # Price momentum
-                features['price_change'] = data['close'].pct_change()
-                features['price_volatility'] = data['close'].rolling(window=20).std()
-            
-            if 'volume' in data.columns:
-                # Volume features
-                features['volume_sma'] = data['volume'].rolling(window=20).mean()
-                features['volume_ratio'] = data['volume'] / features['volume_sma']
-            
-            return pd.DataFrame(features, index=data.index)
-            
-        except Exception as e:
-            tprint_error(f"❌ Fallback feature generation failed: {e}")
-            return pd.DataFrame(index=data.index)
-
-    def _categorize_features(self, feature_names: List[str]) -> List[str]:
-        """Categorize features based on their names."""
-        categories = set()
-        
-        for name in feature_names:
-            name_lower = name.lower()
-            if any(term in name_lower for term in ['sma', 'ema', 'ma']):
-                categories.add('moving_averages')
-            elif any(term in name_lower for term in ['rsi', 'macd', 'stoch']):
-                categories.add('technical_indicators')
-            elif any(term in name_lower for term in ['vol', 'volume']):
-                categories.add('volume_features')
-            elif any(term in name_lower for term in ['volatility', 'std', 'var']):
-                categories.add('volatility_features')
-            elif any(term in name_lower for term in ['change', 'return', 'pct']):
-                categories.add('price_features')
-            else:
-                categories.add('other_features')
-        
-        return list(categories)
-
-    async def _cleanup_m1_resources(self):
-        """Cleanup M1 resources after processing."""
-        try:
-            if self.memory_manager:
-                self.memory_manager.cleanup_all()
-            
-            if self.unified_memory_manager:
-                self.unified_memory_manager.cleanup()
-            
-            force_cleanup()
-            tprint_info("🧹 M1 resources cleaned up")
-            
-        except Exception as e:
-            tprint_warning(f"⚠️ M1 cleanup failed: {e}")
-
-    async def _generate_m1_optimization_report(self, generation_result: M1FeatureGenerationResult, 
-                                             symbol: str, timeframe: str, direction: str, 
-                                             exchange: str = "binance") -> str:
-        """Generate M1 optimization report."""
-        try:
-            from datetime import datetime
-            import os
-            
-            # Create outcomes directory if it doesn't exist
-            outcomes_dir = "outcomes"
-            os.makedirs(outcomes_dir, exist_ok=True)
-            
-            # Generate timestamp for filename
-            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-            report_filename = f"m1_optimized_feature_generation_report_{symbol}_{timeframe}_{direction}_{timestamp}.md"
-            report_path = os.path.join(outcomes_dir, report_filename)
-            
-            # Generate report content
-            report_content = f"""# M1-Optimized Feature Generation Report
-
-## Summary
-- **Symbol**: {symbol}
-- **Exchange**: {exchange}
-- **Timeframe**: {timeframe}
-- **Direction**: {direction}
-- **Generated At**: {datetime.now().strftime("%Y-%m-%d %H:%M:%S")}
-- **Status**: {'✅ SUCCESS' if generation_result.success else '❌ FAILED'}
-
-## M1 Hardware Utilization
-- **Neural Engine Utilization**: {generation_result.neural_engine_utilization:.1f}%
-- **GPU Utilization**: {generation_result.gpu_utilization:.1f}%
-- **CPU Utilization**: {generation_result.cpu_utilization:.1f}%
-- **Memory Efficiency**: {generation_result.memory_efficiency:.1f}%
-
-## Feature Generation Results
-- **Total Features Generated**: {generation_result.n_features_generated}
-- **Generation Time**: {generation_result.generation_time:.3f} seconds
-- **Memory Usage**: {generation_result.memory_usage_mb:.2f} MB
-- **Cache Hit**: {'Yes' if generation_result.cache_hit else 'No'}
-
-## M1 Optimizations Applied
-"""
-            
-            for optimization in generation_result.optimization_applied:
-                report_content += f"- ✅ {optimization}\n"
-            
-            report_content += f"""
-## Feature Categories
-"""
-            
-            for category in generation_result.feature_categories:
-                report_content += f"- **{category.title()}**: Generated\n"
-            
-            report_content += f"""
-## Performance Metrics
-- **Total Processing Time**: {self.performance_stats['total_processing_time']:.2f} seconds
-- **Neural Engine Operations**: {self.performance_stats['neural_engine_operations']}
-- **GPU Accelerations**: {self.performance_stats['gpu_accelerations']}
-- **CPU Optimizations**: {self.performance_stats['cpu_optimizations']}
-- **Memory Optimizations**: {self.performance_stats['memory_optimizations']}
-- **Cache Hits**: {self.performance_stats['cache_hits']}
-- **Memory Savings**: {self.performance_stats['memory_savings_mb']:.2f} MB
-
-## Recommendations
-- ✅ M1 hardware optimization completed successfully
-- 🧠 Neural Engine utilized for ML feature generation
-- 🎮 GPU acceleration applied for vectorized operations
-- 💻 CPU optimization used for traditional features
-- 🧠 Unified memory management applied
-- 📊 Memory efficiency achieved: {generation_result.memory_efficiency:.1f}%
-"""
-            
-            if generation_result.error_message:
-                report_content += f"- **Error**: {generation_result.error_message}\n"
-            
-            # Write report to file
-            with open(report_path, 'w', encoding='utf-8') as f:
-                f.write(report_content)
-            
-            return report_path
-            
-        except Exception as e:
-            self.logger.error(f"Failed to generate M1 optimization report: {e}")
-            return ""
-
-    def _serialize_config(self, config: Any) -> Dict[str, Any]:
-        """Serialize configuration object to dictionary."""
-        try:
-            if hasattr(config, '__dict__'):
-                return {k: v for k, v in config.__dict__.items() if not k.startswith('_')}
-            elif isinstance(config, dict):
-                return config
-            else:
-                return {'config': str(config)}
-        except Exception:
-            return {'config': str(config)}
-
-
-async def handle_m1_optimized_feature_generation_step(
-    symbol: str = "ETHUSDT",
-    timeframe: str = "15m",
-    exchange: str = "binance",
-    direction: str = "longs",
-    intensity: str = None,
-    lookback_days: int = None,
-    start_date: str = None,
-    end_date: str = None,
-    custom_overrides: dict = None,
-    **kwargs
-) -> M1FeatureGenerationResult:
-    """
-    Handler function for M1-optimized feature generation step.
-
-    Args:
-        symbol: Trading symbol (e.g., "ETHUSDT")
-        timeframe: Timeframe (e.g., "15m")
-        exchange: Exchange name (e.g., "binance")
-        direction: Trading direction (e.g., "longs")
-        intensity: Intensity level (e.g., "light", "full", "blank") or None for default
-        lookback_days: Number of days to look back
-        start_date: Start date for data
-        end_date: End date for data
-        custom_overrides: Custom configuration overrides
-        **kwargs: Additional arguments
-
-    Returns:
-        M1FeatureGenerationResult: Result of the M1-optimized feature generation step
-    """
-    # Handle None intensity by defaulting to light mode
-    if intensity is None:
-        intensity = "light"
-
-    try:
-        # Create the step instance
-        step = M1OptimizedFeatureGenerationStep(
-            name="m1_optimized_feature_generation_step",
-            config={
-                'symbol': symbol,
-                'timeframe': timeframe,
-                'exchange': exchange,
-                'direction': direction,
-                'intensity': intensity,
-                'lookback_days': lookback_days,
-                'start_date': start_date,
-                'end_date': end_date,
-                'custom_overrides': custom_overrides
-            }
-        )
-
-        # Create training input
-        training_input = {
-            'symbol': symbol,
-            'timeframe': timeframe,
-            'exchange': exchange,
-            'direction': direction,
-            'intensity': intensity,
-            'lookback_days': lookback_days,
-            'start_date': start_date,
-            'end_date': end_date,
-            'custom_overrides': custom_overrides
-        }
-
-        # Execute the step
-        result = await step.execute(
-            training_input=training_input,
-            pipeline_state={},
-            **kwargs
-        )
-
-        return result
-
-    except Exception as e:
-        import logging
-        logger = logging.getLogger(__name__)
-        logger.error(f"❌ M1-optimized handler function failed: {e}")
-        return M1FeatureGenerationResult(
-            feature_names=[],
-            feature_data=pd.DataFrame(),
-            generated_features=pd.DataFrame(),
-            feature_categories=[],
-            generation_time=0.0,
-            n_features_generated=0,
-            cache_hit=False,
-            memory_usage_mb=0.0,
-            success=False,
-            error_message=str(e),
-            metadata={},
-            optimization_stats={}
-        )
